restructured project packages to devide between responsibilities
