+++ src/test/java/com/github/saschawiegleb/ek/DefaultConfiguration.java
@@ -1,0 +1,7 @@
+package com.github.saschawiegleb.ek;
+
+import com.github.saschawiegleb.ek.entity.Configuration;
+
+public interface DefaultConfiguration {
+    Configuration defaultConfiguration = Configuration.defaults();
+}
