--- src/test/java/com/github/saschawiegleb/ek/api/DefaultConfiguration.java
@@ -1,5 +1,0 @@
-package com.github.saschawiegleb.ek.api;
-
-interface DefaultConfiguration {
-    Configuration defaultConfiguration = Configuration.defaults();
-}
