+++ src/test/java/com/github/saschawiegleb/ek/entity/CategoryTest.java
@@ -1,4 +1,4 @@
-package com.github.saschawiegleb.ek.api;
+package com.github.saschawiegleb.ek.entity;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,9 +6,12 @@ import java.util.NoSuchElementException;
 
 import org.junit.Test;
 
+import com.github.saschawiegleb.ek.DefaultConfiguration;
+import com.github.saschawiegleb.ek.entity.Category;
+
 import javaslang.control.Try;
 
-public class ConfigurationTest implements DefaultConfiguration {
+public class CategoryTest implements DefaultConfiguration {
 
     @Test
     public void testDefaults_resolveCategory_failure() {
